requirements.txt
# pip install -qr requirements.txt

# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow~=9.0.0
PyYAML>=5.3.1
scipy>=1.4.1
tensorboard>=1.5
# torch==1.7.0
# torchvision==0.8.1
tqdm>=4.41.0

# plotting ------------------------------------
seaborn>=0.11.0
pandas~=1.3.5

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
# pycocotools==2.0.0  # COCO mAP

torch~=1.10.1+cu113
torchvision~=0.11.2+cu113
pefile~=2021.9.3
pycocotools~=2.0
albumentations~=1.1.0